Support transparency; many other features (#11)
* Add transparency/devtools flags

Transparency is disabled by default because on osx it requires private APIs. That unfortunately means that an app with transparency *can't* be published to the app store.

Devtools are enabled in all debug builds but are disabled in release builds. Again, on OSX it uses private APIs so an app w/ devtools enabled can't be published to the app store.

* Use transparent flag in build

* Features out the wazoo